# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP Basic with the subscription token as the password. The username is ignored — tokens are self-identifying. This is the form `git clone http://user:\u003ctoken\u003e@host/git/repo` produces, so git-over-HTTP works out of the box.",
        "title": "GiGot API",
        "contact": {
            "name": "Peter van de Pas"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:3417",
    "basePath": "/api",
    "paths": {
        "/admin/accounts": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "GET lists every known account (admins and regulars);\nPOST creates one. Session-cookie authenticated. See\ndocs/design/accounts.md for the data model.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage accounts (admin only)",
                "parameters": [
                    {
                        "description": "Create body (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET response",
                        "schema": {
                            "$ref": "#/definitions/server.AccountListResponse"
                        }
                    },
                    "201": {
                        "description": "POST response",
                        "schema": {
                            "$ref": "#/definitions/server.AccountView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Account already exists",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "GET lists every known account (admins and regulars);\nPOST creates one. Session-cookie authenticated. See\ndocs/design/accounts.md for the data model.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage accounts (admin only)",
                "parameters": [
                    {
                        "description": "Create body (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET response",
                        "schema": {
                            "$ref": "#/definitions/server.AccountListResponse"
                        }
                    },
                    "201": {
                        "description": "POST response",
                        "schema": {
                            "$ref": "#/definitions/server.AccountView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Account already exists",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/accounts/{provider}/{identifier}": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "PATCH updates role, display_name, and/or password\n(local only). DELETE removes the account; the server\nrefuses to remove the last admin so the console can't\nlock itself out. Session-cookie authenticated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage one account by (provider, identifier) (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account identifier",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patch body (PATCH)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.UpdateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.AccountView"
                        }
                    },
                    "204": {
                        "description": "DELETE response"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Would remove the last admin",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "PATCH updates role, display_name, and/or password\n(local only). DELETE removes the account; the server\nrefuses to remove the last admin so the console can't\nlock itself out. Session-cookie authenticated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage one account by (provider, identifier) (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account identifier",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patch body (PATCH)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.UpdateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.AccountView"
                        }
                    },
                    "204": {
                        "description": "DELETE response"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Would remove the last admin",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/accounts/{provider}/{identifier}/tags": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "GET returns the direct tag list; PUT replaces it.\nAccount tags propagate to every subscription the\naccount holds (effective_tags on the token list\nresponses unions account + repo + sub tags).\nCatalogue + assignment diffs land in the system\naudit log. Session-cookie authenticated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage tags on one account (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account identifier",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PUT body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.SetEntityTagsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.EntityTagsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "GET returns the direct tag list; PUT replaces it.\nAccount tags propagate to every subscription the\naccount holds (effective_tags on the token list\nresponses unions account + repo + sub tags).\nCatalogue + assignment diffs land in the system\naudit log. Session-cookie authenticated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Manage tags on one account (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account identifier",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PUT body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.SetEntityTagsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.EntityTagsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.AccountListResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.AccountView"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "server.AccountView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "has_password": {
                    "type": "boolean"
                },
                "identifier": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "subscription_count": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "server.CreateAccountRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "server.EntityTagsResponse": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "repository not found"
                }
            }
        },
...
        "server.SetEntityTagsRequest": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "server.UpdateAccountRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Enter your bearer token as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie minted by POST /api/admin/login (or any successful OAuth callback). Required by every /api/admin/* endpoint and the /fragments/* template server. Browsers attach it automatically; programmatic callers must include the cookie header on every request.",
            "type": "apiKey",
            "name": "gigot_session",
            "in": "cookie"
        }
    }
}